ça exec, froidement.

+++ minishell_m/exec/exec_redir.c
@@ -7,6 +7,8 @@ void here_doc(t_cmd *cmd, char *lim, int i, int *in)
 	(*in)++;
 	gnl = get_next_line(0, 0);
 	cmd->redir->fd[i] = open("/var/tmp/temp", O_RDWR | O_TRUNC | O_CREAT, 0644);
+	if (cmd->redir->fd[i] < 0)
+		return ((*in) = -1000, (void)1);
 	while (ft_findstr(lim, gnl))
 	{
 		write(cmd->redir->fd[i], gnl, ft_strlen(gnl));
@@ -14,7 +16,11 @@ void here_doc(t_cmd *cmd, char *lim, int i, int *in)
 		gnl = get_next_line(0, 0);
 	}
 	cmd->redir->fd[i] = open("/var/tmp/temp", O_RDWR, 0644);
-	dup2(cmd->redir->fd[i], 0);
+	if (cmd->redir->fd[i] < 0)
+		return ((*in) = -1000, (void)1);
+
+	if (dup2(cmd->redir->fd[i], 0) < 0)
+		return ((*in) = -1000, (void)1);
 	close(cmd->redir->fd[i]);
 	unlink("/var/tmp/temp");
 }
@@ -24,12 +30,15 @@ void redir_out(t_cmd *cmd, char *file, int i, int *out)
 	(*out)++;
 	cmd->redir->fd[i] = open(file, O_WRONLY | O_TRUNC | O_CREAT, 0644);
 	if (access(file, W_OK) < 0 && access(file, F_OK) == 0)
-		ft_printf("%s: Permission denied\n", cmd->redir->fd[i]);
+		return (ft_printf_error("%s: Permission denied\n", file),
+			(*out) = -1000, cmd->return_value = 1, (void)1);
 	else if (cmd->redir->fd[i] < 0)
-		ft_printf("%s: a problem has occured with opening outfile\n", cmd->redir->fd[i]);
+		return (ft_printf_error("%s: a problem has occured with opening outfile\n", file),
+			(*out) = -1000, (void)1);
 	else
 	{
-		dup2(cmd->redir->fd[i], 1);
+		if (dup2(cmd->redir->fd[i], 1) < 0)
+			return ((*out) = -1000, (void)1);
 		close(cmd->redir->fd[i]);
 	}
 }
@@ -39,12 +48,15 @@ void append(t_cmd *cmd, char *file, int i, int *out)
 	(*out)++;
 	cmd->redir->fd[i] = open(file, O_WRONLY | O_APPEND | O_CREAT, 0644);
 	if (access(file, W_OK) < 0 && access(file, F_OK) == 0)
-		ft_printf("%s: Permission denied\n", cmd->redir->fd[i]);
+		return (ft_printf_error("%s: Permission denied\n", file),
+			cmd->return_value = 1, (*out) = -1000, (void)1);
 	else if (cmd->redir->fd[i] < 0)
-		ft_printf("%s: a problem has occured with opening outfile\n", cmd->redir->fd[i]);
+		return (ft_printf_error("%s: a problem has occured with opening outfile\n",
+			file),  (*out) = -1000, (void)1);
 	else
 	{
-		dup2(cmd->redir->fd[i], 1);
+		if (dup2(cmd->redir->fd[i], 1) < 0)
+			return ((*out) = -1000, (void)1);
 		close(cmd->redir->fd[i]);
 	}
 }
@@ -54,12 +66,15 @@ void redir_in(t_cmd *cmd, char *file, int i, int *in)
 	(*in)++;
 	cmd->redir->fd[i] = open(file, O_RDONLY);
 	if ((access(file, R_OK) < 0 && access(file, F_OK) == 0))
-		ft_printf("%s: Permission denied\n", file);
+		return (ft_printf_error("%s: Permission denied\n", file),
+			cmd->return_value = 1, (*in) = -1000, (void)1);
 	else if (cmd->redir->fd[i] < 0)
-		ft_printf("%s: no such file or directory\n", file);
+		return (ft_printf_error("%s: no such file or directory\n", file),
+			cmd->return_value = 127, (*in) = -1000, (void)1);
 	else
 	{
-		dup2(cmd->redir->fd[i], 0);
+		if (dup2(cmd->redir->fd[i], 0) < 0)
+			return ((*in) = -1000, (void)1);
 		close(cmd->redir->fd[i]);
 	}
 }
@@ -92,13 +107,10 @@ void pipe_redirection(t_cmd *cmd, int **fds, int in, int out)
 
 void open_redir(t_cmd *cmd, int **fds)
 {
-	int i;
-	int in;
-	int out;
+	int (i) = 0;
+	int (in) = 0;
+	int (out) = 0;
 
-	i = 0;
-	in = 0;
-	out = 0;
 	if (cmd->redir->target)
 	{
 		while (cmd->redir->target[i])
@@ -115,8 +127,9 @@ void open_redir(t_cmd *cmd, int **fds)
 			if (cmd->redir->redir_type[i] == REDIR_IN)
 				redir_in(cmd, cmd->redir->target[i], i, &in);
 			i++;
+			if (in < 0 || out < 0)
+				free_ms(NULL, cmd, cmd->return_value, fds);
 		}
 	}
-	// printf("PROCESS : %d - IN = %d OUT = %d\n", cmd->i, in, out);
 	pipe_redirection(cmd, fds, in, out);
 }
